Fixed EBR.cs

+++ source/Cosmos.HAL2/BlockDevice/EBR.cs
@@ -30,7 +30,7 @@ namespace Cosmos.HAL.BlockDevice
             ParsePartition(aEBR, 462);
         }
 
-        protected void ParsePartition(byte[] aEBR, int aLoc)
+        protected void ParsePartition(byte[] aEBR, ulong aLoc)
         {
             byte xSystemID = aEBR[aLoc + 4];
             // SystemID = 0 means no partition
